Entidade Parceiro

+++ src/main/java/br/gov/sp/fatec/saloon/model/entity/Alugavel.java
@@ -33,10 +33,10 @@ public class Alugavel extends GeneratorId {
 
     // CONSTRUTORES
     public Alugavel(){}
-    public Alugavel( String descr
-                   , String endereco
-                   , int    capacidade
-                   , float  valor){
+    public Alugavel( String     descr
+                   , String     endereco
+                   , int        capacidade
+                   , BigDecimal valor){
 
     }
 
@@ -47,8 +47,8 @@ public class Alugavel extends GeneratorId {
     public void setEndereco(String endereco)                { this.endereco = Texto.left(endereco,500); }
     public int getCapacidade()                              { return capacidade;                        }
     public void setCapacidade(int capacidade)               { this.capacidade = capacidade;             }
-    public float getValor()                                 { return valor;                             }
-    public void setValor(float valor)                       { this.valor = valor;                       }
+    public BigDecimal getValor()                            { return valor;                             }
+    public void setValor(BigDecimal valor)                  { this.valor = valor;                       }
     public AlugavelTipo getAlugavelTipo()                   { return alugavelTipo;                      }
     public void setAlugavelTipo(AlugavelTipo alugavelTipo)  { this.alugavelTipo = alugavelTipo;         }
     public Proprietario getProprietario()                   { return proprietario;                      }
